requirements.txt
streamlit
sentence-transformers
torch
scikit-learn
google-generativeai